refactor(frontend): split PeopleManagement into hooks + sub-components (ATT-292)
Break the 729-line PeopleManagement/index.tsx into focused modules:

- types.ts: PersonRow, AddMode, FilterMode
- usePeopleRows: data fetching + merge into combined rows
- usePeopleMutations: grant/revoke mutations + per-row pending state + toasts
- PeopleHeader: add-person dropdown/button
- PeopleTable + PeopleRowActions: table rendering + row action buttons
- AddPersonModal / IntroductionCommentModal / HistoryModalLoader: modals

index.tsx becomes a thin orchestrator that wires hooks to sub-components.
No behavior change.
